Move utils/pooling to training/pooling.py
