+++ recipes/ppo_full_finetune_single_device.py
@@ -741,7 +741,7 @@ class PPOFullFinetuneRecipeSingleDevice(FTRecipeInterface):
 
         # step 5.1 the scores from the reward model are the logits for the last non-padding token in
         # each (query, truncated-response) pair
-        seq_lens = utils.get_unmasked_sequence_lengths(response_padding_masks)
+        seq_lens = training.get_unmasked_sequence_lengths(response_padding_masks)
         scores = scores[torch.arange(batch_size), seq_lens + context_length].squeeze(-1)
 
         # step 5.2 if configured, apply any penalties for sequences without EOS tokens
