+++ torchtune/utils/__init__.py
@@ -35,7 +35,6 @@ from .memory import (  # noqa
     register_optim_in_bwd_hooks,
     set_activation_checkpointing,
 )
-from .pooling import get_unmasked_sequence_lengths
 from .seed import set_seed
 
 __all__ = [
@@ -50,7 +49,6 @@ __all__ = [
     "is_distributed",
     "lora_fsdp_wrap_policy",
     "get_full_finetune_fsdp_wrap_policy",
-    "get_unmasked_sequence_lengths",
     "set_activation_checkpointing",
     "set_seed",
     "torch_version_ge",
